move boolean decomposition to separate file
